airbyte-8174 🎉 Destination BigQuery: Use byte-based buffering batching. Refactoring.
