+++ airbyte-integrations/connectors/destination-bigquery/src/main/java/io/airbyte/integrations/destination/bigquery/BigQueryRecordConsumer.java
@@ -62,15 +62,15 @@ public class BigQueryRecordConsumer extends FailureTrackingAirbyteMessageConsume
   private final Consumer<AirbyteMessage> outputRecordCollector;
   private final boolean isGcsUploadingMode;
   private final boolean isKeepFilesInGcs;
-  private AirbyteMessage lastStateMessage = null;
   private long bufferSizeInBytes;
 
   private final List<AirbyteMessage> buffer;
   private static final int MAX_BATCH_SIZE_BYTES = 1024 * 1024 * 1024 / 4; // 256 mib
   private final ConfiguredAirbyteCatalog catalog;
+
+  private AirbyteMessage lastStateMessage = null;
   private AirbyteMessage lastFlushedState;
   private AirbyteMessage pendingState;
-
 
   public BigQueryRecordConsumer(final BigQuery bigquery,
       final Map<AirbyteStreamNameNamespacePair, BigQueryWriteConfig> writeConfigs,
@@ -106,35 +106,32 @@ public class BigQueryRecordConsumer extends FailureTrackingAirbyteMessageConsume
   }
 
   private void processRecord(AirbyteMessage message) {
-    final AirbyteRecordMessage recordMessage = message.getRecord();
-
-    // ignore other message types.
-    final AirbyteStreamNameNamespacePair pair = AirbyteStreamNameNamespacePair.fromRecordMessage(recordMessage);
-    if (!writeConfigs.containsKey(pair)) {
-      throw new IllegalArgumentException(
-          String.format("Message contained record from a stream that was not in the catalog. \ncatalog: %s , \nmessage: %s",
-              Jsons.serialize(catalog), Jsons.serialize(recordMessage)));
-    }
-    final BigQueryWriteConfig writer = writeConfigs.get(pair);
-
-    long messageSizeInBytes = ByteUtils.getSizeInBytes(Jsons.serialize(recordMessage.getData()));
+    long messageSizeInBytes = ByteUtils.getSizeInBytes(Jsons.serialize(message));
     if (bufferSizeInBytes + messageSizeInBytes >= MAX_BATCH_SIZE_BYTES) {
       // select the way of uploading - normal or through the GCS
-      if (writer.getGcsCsvWriter() != null) {
-        flushQueueToDestination(writer, new BigQueryUploadGCSStrategy());
-      } else {
-        flushQueueToDestination(writer, new BigQueryUploadBasicStrategy());
-      }
+      flushQueueToDestination();
       bufferSizeInBytes = 0;
     }
     buffer.add(message);
     bufferSizeInBytes += messageSizeInBytes;
   }
 
-  protected void flushQueueToDestination(BigQueryWriteConfig writer, BigQueryUploadStrategy bigQueryUploadStrategy) {
-    for (final AirbyteMessage airbyteMessage : buffer) {
-      bigQueryUploadStrategy.upload(writer, airbyteMessage, catalog);
-    }
+  protected void flushQueueToDestination() {
+    // ignore other message types.
+    buffer.forEach(airbyteMessage -> {
+      final AirbyteStreamNameNamespacePair pair = AirbyteStreamNameNamespacePair.fromRecordMessage(airbyteMessage.getRecord());
+      if (!writeConfigs.containsKey(pair)) {
+        throw new IllegalArgumentException(
+            String.format("Message contained record from a stream that was not in the catalog. \ncatalog: %s , \nmessage: %s",
+                Jsons.serialize(catalog), Jsons.serialize(airbyteMessage.getRecord())));
+      }
+      final BigQueryWriteConfig writer = writeConfigs.get(pair);
+      if (writer.getGcsCsvWriter() != null) {
+        new BigQueryUploadGCSStrategy(bigquery).upload(writer, airbyteMessage, catalog);
+      } else {
+        new BigQueryUploadBasicStrategy(bigquery, catalog, outputRecordCollector, lastStateMessage);
+      }
+    });
 
     buffer.clear();
 
@@ -147,12 +144,14 @@ public class BigQueryRecordConsumer extends FailureTrackingAirbyteMessageConsume
   @Override
   public void close(final boolean hasFailed) {
     LOGGER.info("Started closing all connections");
+
     // process gcs streams
     if (isGcsUploadingMode) {
-      closeGcsStreamsAndCopyDataToBigQuery(hasFailed);
+      final List<BigQueryWriteConfig> gcsWritersList = writeConfigs.values().parallelStream()
+          .filter(el -> el.getGcsCsvWriter() != null)
+          .collect(Collectors.toList());
+      new BigQueryUploadGCSStrategy(bigquery).close(gcsWritersList, hasFailed);
     }
-
-    closeNormalBigqueryStreams(hasFailed);
 
     if (isGcsUploadingMode && !isKeepFilesInGcs) {
       deleteDataFromGcsBucket();
@@ -160,148 +159,6 @@ public class BigQueryRecordConsumer extends FailureTrackingAirbyteMessageConsume
 
     if (lastFlushedState != null) {
       outputRecordCollector.accept(lastFlushedState);
-    }
-  }
-
-  private void closeGcsStreamsAndCopyDataToBigQuery(final boolean hasFailed) {
-    final List<BigQueryWriteConfig> gcsWritersList = writeConfigs.values().parallelStream()
-        .filter(el -> el.getGcsCsvWriter() != null)
-        .collect(Collectors.toList());
-
-    if (!gcsWritersList.isEmpty()) {
-      LOGGER.info("GCS connectors that need to be closed:" + gcsWritersList);
-      gcsWritersList.parallelStream().forEach(writer -> {
-        final GcsCsvWriter gcsCsvWriter = writer.getGcsCsvWriter();
-
-        try {
-          LOGGER.info("Closing connector:" + gcsCsvWriter);
-          gcsCsvWriter.close(hasFailed);
-        } catch (final IOException | RuntimeException e) {
-          LOGGER.error(String.format("Failed to close %s gcsWriter, \n details: %s", gcsCsvWriter, e.getMessage()));
-          printHeapMemoryConsumption();
-          throw new RuntimeException(e);
-        }
-      });
-    }
-
-    // copy data from tmp gcs storage to bigquery tables
-    writeConfigs.values().stream()
-        .filter(pair -> pair.getGcsCsvWriter() != null)
-        .forEach(pair -> {
-          try {
-            loadCsvFromGcsTruncate(pair);
-          } catch (final Exception e) {
-            LOGGER.error("Failed to load data from GCS CSV file to BigQuery tmp table with reason: " + e.getMessage());
-            throw new RuntimeException(e);
-          }
-        });
-  }
-
-  private void loadCsvFromGcsTruncate(final BigQueryWriteConfig bigQueryWriteConfig)
-      throws Exception {
-    try {
-
-      final TableId tmpTable = bigQueryWriteConfig.getTmpTable();
-      final Schema schema = bigQueryWriteConfig.getSchema();
-
-      final String csvFile = bigQueryWriteConfig.getGcsCsvWriter().getGcsCsvFileLocation();
-
-      // Initialize client that will be used to send requests. This client only needs to be created
-      // once, and can be reused for multiple requests.
-      LOGGER.info(String.format("Started copying data from %s GCS csv file to %s tmp BigQuery table with schema: \n %s",
-          csvFile, tmpTable, schema));
-
-      final CsvOptions csvOptions = CsvOptions.newBuilder().setEncoding(UTF8).setSkipLeadingRows(1).build();
-
-      final LoadJobConfiguration configuration =
-          LoadJobConfiguration.builder(tmpTable, csvFile)
-              .setFormatOptions(csvOptions)
-              .setSchema(schema)
-              .setWriteDisposition(bigQueryWriteConfig.getSyncMode())
-              .build();
-
-      // For more information on Job see:
-      // https://googleapis.dev/java/google-cloud-clients/latest/index.html?com/google/cloud/bigquery/package-summary.html
-      // Load the table
-      final Job loadJob = bigquery.create(JobInfo.of(configuration));
-
-      LOGGER.info("Created a new job GCS csv file to tmp BigQuery table: " + loadJob);
-      LOGGER.info("Waiting for job to complete...");
-
-      // Load data from a GCS parquet file into the table
-      // Blocks until this load table job completes its execution, either failing or succeeding.
-      final Job completedJob = loadJob.waitFor();
-
-      // Check for errors
-      if (completedJob == null) {
-        LOGGER.error("Job not executed since it no longer exists.");
-        throw new Exception("Job not executed since it no longer exists.");
-      } else if (completedJob.getStatus().getError() != null) {
-        // You can also look at queryJob.getStatus().getExecutionErrors() for all
-        // errors, not just the latest one.
-        final String msg = "BigQuery was unable to load into the table due to an error: \n"
-            + loadJob.getStatus().getError();
-        LOGGER.error(msg);
-        throw new Exception(msg);
-      }
-      LOGGER.info("Table is successfully overwritten by CSV file loaded from GCS");
-    } catch (final BigQueryException | InterruptedException e) {
-      LOGGER.error("Column not added during load append \n" + e.toString());
-      throw new RuntimeException("Column not added during load append \n" + e.toString());
-    }
-  }
-
-  private void closeNormalBigqueryStreams(final boolean hasFailed) {
-    try {
-      writeConfigs.values().parallelStream().forEach(bigQueryWriteConfig -> Exceptions.toRuntime(() -> {
-        final TableDataWriteChannel writer = bigQueryWriteConfig.getWriter();
-        try {
-          writer.close();
-        } catch (final IOException | RuntimeException e) {
-          LOGGER.error(String.format("Failed to close writer: %s, \nStreams numbers: %s",
-              writer.getJob(), catalog.getStreams().size()));
-          printHeapMemoryConsumption();
-          throw new RuntimeException(e);
-        }
-      }));
-
-      LOGGER.info("Waiting for jobs to be finished/closed");
-      writeConfigs.values().forEach(bigQueryWriteConfig -> Exceptions.toRuntime(() -> {
-        if (bigQueryWriteConfig.getWriter().getJob() != null) {
-          try {
-            bigQueryWriteConfig.getWriter().getJob().waitFor();
-          } catch (final RuntimeException e) {
-            LOGGER.error(
-                String.format("Failed to process a message for job: %s, \nStreams numbers: %s, \nSyncMode: %s, \nTableName: %s, \nTmpTableName: %s",
-                    bigQueryWriteConfig.getWriter().getJob(), catalog.getStreams().size(), bigQueryWriteConfig.getSyncMode(),
-                    bigQueryWriteConfig.getTable(), bigQueryWriteConfig.getTmpTable()));
-            printHeapMemoryConsumption();
-            throw new RuntimeException(e);
-          }
-        }
-      }));
-
-      if (!hasFailed) {
-        LOGGER.info("Replication finished with no explicit errors. Copying data from tmp tables to permanent");
-        writeConfigs.values()
-            .forEach(
-                bigQueryWriteConfig -> {
-                  if (bigQueryWriteConfig.getSyncMode().equals(WriteDisposition.WRITE_APPEND)) {
-                    partitionIfUnpartitioned(bigQueryWriteConfig, bigquery, bigQueryWriteConfig.getTable());
-                  }
-                  copyTable(bigquery, bigQueryWriteConfig.getTmpTable(), bigQueryWriteConfig.getTable(),
-                      bigQueryWriteConfig.getSyncMode());
-                });
-        // BQ is still all or nothing if a failure happens in the destination.
-        outputRecordCollector.accept(lastStateMessage);
-      } else {
-        LOGGER.warn("Had errors while replicating");
-      }
-    } finally {
-      // clean up tmp tables;
-      LOGGER.info("Removing tmp tables...");
-      writeConfigs.values().forEach(bigQueryWriteConfig -> bigquery.delete(bigQueryWriteConfig.getTmpTable()));
-      LOGGER.info("Finishing destination process...completed");
     }
   }
 
@@ -331,85 +188,5 @@ public class BigQueryRecordConsumer extends FailureTrackingAirbyteMessageConsume
       }
       s3Client.shutdown();
     });
-  }
-
-  // https://cloud.google.com/bigquery/docs/managing-tables#copying_a_single_source_table
-  private static void copyTable(
-                                final BigQuery bigquery,
-                                final TableId sourceTableId,
-                                final TableId destinationTableId,
-                                final WriteDisposition syncMode) {
-    final CopyJobConfiguration configuration = CopyJobConfiguration.newBuilder(destinationTableId, sourceTableId)
-        .setCreateDisposition(CreateDisposition.CREATE_IF_NEEDED)
-        .setWriteDisposition(syncMode)
-        .build();
-
-    final Job job = bigquery.create(JobInfo.of(configuration));
-    final ImmutablePair<Job, String> jobStringImmutablePair = BigQueryUtils.executeQuery(job);
-    if (jobStringImmutablePair.getRight() != null) {
-      LOGGER.error("Failed on copy tables with error:" + job.getStatus());
-      throw new RuntimeException("BigQuery was unable to copy table due to an error: \n" + job.getStatus().getError());
-    }
-    LOGGER.info("successfully copied table: {} to table: {}", sourceTableId, destinationTableId);
-  }
-
-  private void partitionIfUnpartitioned(final BigQueryWriteConfig bigQueryWriteConfig,
-                                        final BigQuery bigquery,
-                                        final TableId destinationTableId) {
-    try {
-      final QueryJobConfiguration queryConfig = QueryJobConfiguration
-          .newBuilder(
-              String.format("SELECT max(is_partitioning_column) as is_partitioned FROM `%s.%s.INFORMATION_SCHEMA.COLUMNS` WHERE TABLE_NAME = '%s';",
-                  bigquery.getOptions().getProjectId(),
-                  destinationTableId.getDataset(),
-                  destinationTableId.getTable()))
-          .setUseLegacySql(false)
-          .build();
-      final ImmutablePair<Job, String> result = BigQueryUtils.executeQuery(bigquery, queryConfig);
-      result.getLeft().getQueryResults().getValues().forEach(row -> {
-        if (!row.get("is_partitioned").isNull() && row.get("is_partitioned").getStringValue().equals("NO")) {
-          LOGGER.info("Partitioning existing destination table {}", destinationTableId);
-          final String tmpPartitionTable = Strings.addRandomSuffix("_airbyte_partitioned_table", "_", 5);
-          final TableId tmpPartitionTableId = TableId.of(destinationTableId.getDataset(), tmpPartitionTable);
-          // make sure tmpPartitionTable does not already exist
-          bigquery.delete(tmpPartitionTableId);
-          // Use BigQuery SQL to copy because java api copy jobs does not support creating a table from a
-          // select query, see:
-          // https://cloud.google.com/bigquery/docs/creating-partitioned-tables#create_a_partitioned_table_from_a_query_result
-          final QueryJobConfiguration partitionQuery = QueryJobConfiguration
-              .newBuilder(
-                  getCreatePartitionedTableFromSelectQuery(bigQueryWriteConfig.getSchema(), bigquery.getOptions().getProjectId(), destinationTableId,
-                      tmpPartitionTable))
-              .setUseLegacySql(false)
-              .build();
-          BigQueryUtils.executeQuery(bigquery, partitionQuery);
-          // Copying data from a partitioned tmp table into an existing non-partitioned table does not make it
-          // partitioned... thus, we force re-create from scratch by completely deleting and creating new
-          // table.
-          bigquery.delete(destinationTableId);
-          copyTable(bigquery, tmpPartitionTableId, destinationTableId, WriteDisposition.WRITE_EMPTY);
-          bigquery.delete(tmpPartitionTableId);
-        }
-      });
-    } catch (final InterruptedException e) {
-      LOGGER.warn("Had errors while partitioning: ", e);
-    }
-  }
-
-  protected String getCreatePartitionedTableFromSelectQuery(final Schema schema,
-                                                            final String projectId,
-                                                            final TableId destinationTableId,
-                                                            final String tmpPartitionTable) {
-    return String.format("create table `%s.%s.%s` (", projectId, destinationTableId.getDataset(), tmpPartitionTable)
-        + schema.getFields().stream()
-            .map(field -> String.format("%s %s", field.getName(), field.getType()))
-            .collect(Collectors.joining(", "))
-        + ") partition by date("
-        + JavaBaseConstants.COLUMN_NAME_EMITTED_AT
-        + ") as select "
-        + schema.getFields().stream()
-            .map(Field::getName)
-            .collect(Collectors.joining(", "))
-        + String.format(" from `%s.%s.%s`", projectId, destinationTableId.getDataset(), destinationTableId.getTable());
   }
 }
